requirements.txt
numpy
pandas
scikit-learn
scipy
matplotlib
flask
scikit-surprise
tqdm
joblib
